pkg/rand: add implementation for Plan 9
Plan 9 doesn't have /dev/urandom. The crypto/rand package in the
standard library implements its own pseudorandom generator seeded by
/dev/random on Plan 9. We just wrap the crypto/rand implementation,
adding a context.
