+++ pkg/rand/random_unix.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build aix darwin dragonfly freebsd nacl netbsd openbsd plan9 solaris
+// +build aix darwin dragonfly freebsd nacl netbsd openbsd solaris
 
 package rand
